moved tests into individual files, needs more linting fixes
